+++ src/cascadia/TerminalSettingsModel/TerminalSettings.h
@@ -21,7 +21,7 @@ Author(s):
 #include <DefaultSettings.h>
 #include <conattrs.hpp>
 
-using IFontAxesMap = winrt::Windows::Foundation::Collections::IMap<winrt::hstring, int64_t>;
+using IFontAxesMap = winrt::Windows::Foundation::Collections::IMap<winrt::hstring, int32_t>;
 using IFontFeatureMap = winrt::Windows::Foundation::Collections::IMap<winrt::hstring, uint32_t>;
 
 // fwdecl unittest classes
