+++ src/renderer/dx/DxRenderer.hpp
@@ -73,7 +73,7 @@ namespace Microsoft::Console::Render
         HANDLE GetSwapChainHandle();
 
         void SetFontFeatures(const std::unordered_map<std::wstring_view, uint32_t> features);
-        void SetFontAxes(const std::unordered_map<std::wstring_view, int64_t> axes);
+        void SetFontAxes(const std::unordered_map<std::wstring_view, int32_t> axes);
 
         // IRenderEngine Members
         [[nodiscard]] HRESULT Invalidate(const SMALL_RECT* const psrRegion) noexcept override;
